feat: MoveController에 존재하던 db 관련 로직을 서비스에게 위임한다
